requirements.txt
altair==5.0.1
aniso8601==9.0.1
ansi2html==1.8.0
arrow==1.2.3
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
decorator==5.1.1
enum-compat==0.0.3
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvgpu==0.10.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pynvml==11.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
semver==3.0.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
streamlit==1.11.1
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-model-archiver==0.8.1
torchdata==0.6.1
torchserve==0.8.1
torchserve-dashboard==0.6.0
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.6
zipp==3.16.2
accelerate==0.21.0
altair==5.0.1
aniso8601==9.0.1
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
arrow==1.2.3
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
decorator==5.1.1
dnspython==2.4.0
email-validator==2.0.0.post2
enum-compat==0.0.3
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvgpu==0.10.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
pluggy==1.2.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pydantic==2.0.3
pydantic-core==2.3.0
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pynvml==11.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
semver==3.0.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.11.1
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-model-archiver==0.8.1
torchdata==0.6.1
torchserve==0.8.1
torchserve-dashboard==0.6.0
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.16.2
